extract HttpServerEmbeddedServer
